Add device info at the beginning of saved logs file
(cherry picked from commit 85677bb0d72a2d59ed55dbd89cb2954d014fc763)
